tests/wpt/tests/html/semantics/menu/tentative/focus-menu-elements-arrowoperations.html

<!DOCTYPE html>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<link rel=author href=mailto:dizhangg@chromium.org>
<link rel=help href=https://open-ui.org/components/menu.explainer>

<menubar>
 <menuitem id=btn commandfor="list" command="toggle-menu">Open</menuitem>
</menubar>

<menulist id="list">
 <menuitem id="A1">Command A1</menuitem>
 <menuitem id="A2">Command A2</menuitem>
 <menuitem id="A3">Command A3</menuitem>
</menulist>

<menubar id="bar">
  <menuitem id="B1">Command B1</menuitem>
  <menuitem id="B2">Command B2</menuitem>
  <menuitem id="B3">Command B3</menuitem>
</menubar>

<style>
  .nodelay menuitem {
    interest-delay: 0s;
  }
  .longdelay menuitem {
    interest-delay: 1000s;
  }
</style>

<script>

const Enter = '\uE007';
const ArrowLeft = '\uE012';
const ArrowUp = '\uE013';
const ArrowRight = '\uE014';
const ArrowDown = '\uE015';

[false, true].forEach(useInterest => {
  document.body.className = useInterest ? 'nodelay' : 'longdelay';
  const description = useInterest ? ' (no delays)' : ' (long delays)';
  promise_test(async (t) => {
    await test_driver.click(btn);
    assert_equals(document.activeElement, btn);
    await test_driver.send_keys(document.activeElement, Enter);
    assert_equals(document.activeElement, btn, 'btn invoked menulist, but focus is still on btn.');
    await test_driver.send_keys(document.activeElement, ArrowDown);
    assert_equals(document.activeElement, A1, 'arrow down moves focus into menulist.');
    await test_driver.send_keys(document.activeElement, ArrowDown);
    assert_equals(document.activeElement, A2, 'arrow down changes to next menuitem.');
    await test_driver.send_keys(document.activeElement, ArrowUp);
    assert_equals(document.activeElement, A1, 'arrow up changes to previous menuitem.');
    await test_driver.send_keys(document.activeElement, ArrowUp);
    assert_equals(document.activeElement, A3, 'arrow up key loops.');
    await test_driver.send_keys(document.activeElement, ArrowDown);
    assert_equals(document.activeElement, A1, 'arrow down key loops.');

    list.hidePopover();
    await test_driver.click(btn);
    await test_driver.send_keys(document.activeElement, ArrowDown);
    assert_equals(document.activeElement, A1, 'arrow down moves focus into menulist.');
    await test_driver.send_keys(document.activeElement, ArrowLeft);
    assert_equals(document.activeElement, btn, 'arrow left close the menulist and focus on invoker.');

    await test_driver.click(btn);
    await test_driver.send_keys(document.activeElement, ArrowDown);
    assert_equals(document.activeElement, A1, 'arrow down moves focus into menulist.');
    await test_driver.send_keys(document.activeElement, ArrowRight);
    assert_equals(document.activeElement, btn, 'arrow right close the menulist and focus on invoker.');
  }, `Should use arrow keys to move between menuitems in menulist${description}`);

  promise_test(async (t) => {
    list.style.display = 'block';
    await test_driver.click(btn);
    await test_driver.send_keys(document.activeElement, ArrowDown);
    assert_equals(document.activeElement, A1, 'arrow down moves focus into menulist.');
    await test_driver.send_keys(document.activeElement, ArrowDown);
    assert_equals(document.activeElement, A2, 'arrow down changes to next menuitem.');
    await test_driver.send_keys(document.activeElement, ArrowUp);
    assert_equals(document.activeElement, A1, 'arrow up changes to previous menuitem.');
    await test_driver.send_keys(document.activeElement, ArrowLeft);
    assert_equals(document.activeElement, btn, 'arrow left moves focus from menulist to invoker.');
    list.style.display = '';
  }, `Should use arrow keys to move between menuitems in menulist with display block${description}`);

  promise_test(async (t) => {
    await test_driver.click(B1);
    assert_equals(document.activeElement, B1);
    await test_driver.send_keys(document.activeElement, ArrowUp);
    assert_equals(document.activeElement, B1, 'arrow up does not change current focused element.');
    await test_driver.send_keys(document.activeElement, ArrowDown);
    assert_equals(document.activeElement, B1, 'arrow down does not change current focused element.');
    await test_driver.send_keys(document.activeElement, ArrowRight);
    assert_equals(document.activeElement, B2, 'arrow right changes to next menuitem.');
    await test_driver.send_keys(document.activeElement, ArrowLeft);
    assert_equals(document.activeElement, B1, 'arrow left changes to previous menuitem.');
    await test_driver.send_keys(document.activeElement, ArrowLeft);
    assert_equals(document.activeElement, B3, 'arrow left key loops.');
    await test_driver.send_keys(document.activeElement, ArrowRight);
    assert_equals(document.activeElement, B1, 'arrow right key loops.');
  }, `Should use arrow keys to move between menuitems in menubar${description}`);
});

</script>
